requirements.txt
pandas==2.3.2
streamlit==1.49.1
Flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.25
pymysql==1.1.0
pymongo==4.6.1
python-dotenv==1.0.1
streamlit==1.37.0
alembic==1.13.1
